Upgrade Playwright and streamline e2e storage setup
