+++ e2e/tests/generation-flow.spec.ts
@@ -1,6 +1,7 @@
 import { test, expect } from '../fixtures/base';
 import { GenerationPreviewPage } from '../pages/generation-preview.page';
 import { createSettingsStorage } from '../fixtures/test-data/settings';
+import { setSettingsAndGenerationSession } from '../fixtures/browser-storage';
 
 const SETTINGS_STORAGE = createSettingsStorage();
 
@@ -19,13 +20,10 @@ const GENERATION_SESSION = JSON.stringify({
 
 test.describe('Generation Flow', () => {
   test.beforeEach(async ({ page }) => {
-    await page.addInitScript(
-      ({ settings, session }) => {
-        localStorage.setItem('settings-storage', settings);
-        sessionStorage.setItem('generationSession', session);
-      },
-      { settings: SETTINGS_STORAGE, session: GENERATION_SESSION },
-    );
+    await setSettingsAndGenerationSession(page, {
+      settings: SETTINGS_STORAGE,
+      session: GENERATION_SESSION,
+    });
   });
 
   test('completes generation pipeline and redirects to classroom', async ({ page, mockApi }) => {
